Deleting contents of Ported folder
This folder (`Ported`) used to contain a set of samples ported to DOTS. These have been deleted in master so as to avoid confusing end users. Why confusing? Well, the ports were each implemented according to whatever was the best practice and tech at the time. This tends to get out of date quickly, and the ports are not maintained (they are intended as an exercise), so let's just tuck them away in branches.

The ports that used to be here can be found by checking out the `archived-ported-samples` branch.

_When contributing a DOTS-ported sample, please always do so in a branch that includes your username in the branch name._