JobClothSimEcs - different approaches
Break JobClothSimEcs system into different systems, for illustration
of the three types of approaches I tried
* manual iteration with IJob for constraint sim, and IJob for mesh
sim
* manual iteraiton with IJob for constraint sim and IJobParallelBatch
for mesh sim
* ForEach with all processing in one job
Also updated the authoring component to let you select one at runtime.

Implemented an implicitly iterated ForEach system in one pass.  Uses
BlobAsset instead of SharedComponent.

Broke out RenderMesh copying to a separate system.

Also after some profiling I saw that maintaining a separate simGroup
just for the entity count was doubling up in the manual iteration
cases, because I was not keeping the queries identical.  It's not
obvious if it's possible to explicitly run a job with a query, and
it doesn't seem to be possible to build the
ForEachLambdaJobDescription and extract the query before running, so I
punted and just build combine job handles as we go.  Kind of gross but
there you go.

Added an enumerated set of test scenes.

Updated readme a bit.