Fix off by 1 bug that gives invalid model, hangs editor due to CAS bug in hybrid renderer
Using an invalid distance field model would hang the editor. This is because when using an invalid model, GetDistance() in ParticleFixedRateSimulationSystem.cs would return a normal of (0,0,0) and a distance of float.MaxValue. Normalizing the all 0 normal results in NaNs, which propagate to the LocalToWorld matrix via the Euler integration calculation in the ParticleFixedRateSimulation job. The hybrid renderer uses the LocalToWorld matrix to calculate AABBs for rendering. When processing Chunks to submit batches to the SRP batcher, CommitBatchAABB() calculates the min and max points of the batch AABB using AtomicMin/Max, which uses a CAS. Because of the NaNs stored in floats[], the if (currentValue <= value) statement in AtomicMin would never be true, so it would infinitely loop, hanging the editor.

The CAS issue has been addressed with https://github.com/Unity-Technologies/dots/pull/5911