Spawn particles in same distribution as original
Random.NextFloatDirection() gives you a unit length vector in a random direction, so you get all the particles spawing on the skin of a sphere. The original uses Random.insideUnitSphere, which scales the unit vector by the cube root of a random float between 0 and 1 to give an even distribution within the volume of the sphere.