Work on making tornado forces parallel
Build tornado data to parallel writers, set the dispatch to schedule parallel.  This doesn't buy us much yet since we only have 2 tornados that fit into a single chunk, but should be good longterm.

Set the forces update to schedule parallel, this is a big win.  At a 500x500 beam grid with 2 tornados, it was roughly an 11ms savings.

Did some formatting and cleanup while I was in here.