Turns out we had a situation.
When we made TargetValidationSystem.cs, attacking bees wasn't a thing yet. This means that, once we could attack bees, we weren't considering that possibility. This means that, whenever a bee targeted another, validation didn't recognize it and scrapped the targeting, forcing reacquisition.

No more.

Also slowed down the bees when attacking. It makes things a tad too crazy. Need to work on dampening the speeds faster if we're gonna make them go faster so they don't slide all over the place.